(Both) - Fixed qun image posting failed if directory contains non-English characters - Fixed negative temperature shows as 655xx
(MIMQQ3)
- Fixed unable to show weather (always shows not connected to network)

+++ mimqq3/libeva2/evaweather.cpp
@@ -109,9 +109,9 @@ int Weather::parseBean(unsigned char* buf) {
 	delete[] pszTemp;
 	pos+=len;
 
-	lowTemperature = htons(*(short*)(buf+pos));
+	lowTemperature = (short)htons(*(short*)(buf+pos));
 	pos+=2;
-	highTemperature = htons(*(short*)(buf+pos));
+	highTemperature = (short)htons(*(short*)(buf+pos));
 	pos+=2;
 
 	pos++;
